Use RaggedBuffer to fully batch PPO train loop (#40)
* Use RaggedBuffer to Fully batch PPO train loop

* Simplify, add ragged-buffer dependency

* Remove unused imports

* Remove type: ignore

* Add missing dependency

+++ entity_gym/envs/move_to_origin.py
@@ -121,7 +121,8 @@ class MoveToOrigin(Environment):
                             self.y_velocity,
                             self.step,
                         ]
-                    ]
+                    ],
+                    dtype=np.float32,
                 ),
             },
             action_masks={
